Renamed some parameters in API to be more clear - onResponse -> onDone, our fetch chains behave atomically - changed success function type to `() -> Unit`
