+++ Density/app/src/main/AndroidManifest.xml
@@ -33,7 +33,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".facilitydetail.FacilityInfoPage">
+        <activity android:name=".facilityinfo.FacilityInfoPage">
             <intent-filter>
                 <action android:name="android.intent.action.VIEW" />
