+++ Density/app/src/main/res/layout/facility_info_page.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:clickable="true"
     android:focusable="true"
-    tools:context=".facilitydetail.FacilityInfoPage">
+    tools:context=".facilityinfo.FacilityInfoPage">
 
     <LinearLayout
         android:id="@+id/topView"
